+++ service-order/src/main/java/com/wulb2018/order/model/dto/OrderAddDTO.java
@@ -37,11 +37,11 @@ public class OrderAddDTO {
     private Double price;
 
     @ApiModelProperty("委托数量")
-    private Double quantity;
+    private Integer quantity;
 
     @Hidden
     @ApiModelProperty("已成交数量")
-    private Double filledQuantity;
+    private Integer filledQuantity;
 
     @Hidden
     @ApiModelProperty("冻结资金/资产")
